Camera Sensor working Simulation

+++ simulation/revolve2/simulation/scene/_simulation_state.py
@@ -69,7 +69,7 @@ class SimulationState(ABC):
         """
 
     @abstractmethod
-    def get_camera_view(self, camera_sensor: CameraSensor) -> NDArray[np.float_]:
+    def get_camera_view(self, camera_sensor: CameraSensor) -> NDArray[np.uint8]:
         """
         Get the camera view.
